+++ packages/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crossed/demo
 
+## 0.2.1
+
+### Patch Changes
+
+- fix error styled('button')
+
 ## 0.2.0
 
 ### Minor Changes